requirements.txt
Flask
google-genai
requests
python-dotenv